pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "cap-srp"
version = "0.1.0"
description = "CAP-SRP: Cryptographic proof that AI systems refused to generate harmful content"
readme = "README.md"
license = {text = "Apache-2.0"}
requires-python = ">=3.10"
authors = [
    {name = "VeritasChain Standards Organization", email = "info@veritaschain.org"}
]
maintainers = [
    {name = "Tomohiro Kamimura", email = "kamimura@veritaschain.org"}
]
keywords = [
    "ai-safety",
    "cryptography",
    "audit-trail",
    "content-moderation",
    "merkle-tree",
    "transparency",
    "compliance",
    "eu-ai-act",
    "dsa"
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Intended Audience :: Financial and Insurance Industry",
    "Intended Audience :: Legal Industry",
    "License :: OSI Approved :: Apache Software License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Security :: Cryptography",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Topic :: System :: Logging",
    "Typing :: Typed"
]
dependencies = [
    "cryptography>=42.0.0",
    "PyNaCl>=1.5.0",
    "pydantic>=2.5.0",
    "python-dateutil>=2.8.2",
    "streamlit>=1.32.0",
    "plotly>=5.18.0",
    "pandas>=2.1.0",
    "rich>=13.7.0",
    "click>=8.1.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-cov>=4.1.0",
    "pytest-asyncio>=0.23.0",
    "black>=24.1.0",
    "isort>=5.13.0",
    "mypy>=1.8.0",
    "pre-commit>=3.6.0",
    "ruff>=0.2.0",
    "jsonschema>=4.21.0",
]
api = [
    "fastapi>=0.109.0",
    "uvicorn>=0.27.0",
    "httpx>=0.26.0",
]
tsa = [
    "rfc3161ng>=2.1.3",
]
all = [
    "cap-srp[dev,api,tsa]",
]

[project.urls]
Homepage = "https://veritaschain.org"
Documentation = "https://github.com/veritaschain/cap-srp-dashboard#readme"
Repository = "https://github.com/veritaschain/cap-srp-dashboard"
Issues = "https://github.com/veritaschain/cap-srp-dashboard/issues"
Changelog = "https://github.com/veritaschain/cap-srp-dashboard/blob/main/CHANGELOG.md"

[project.scripts]
cap-srp = "cap_srp.cli:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["cap_srp*"]

[tool.setuptools.package-data]
cap_srp = ["py.typed"]

[tool.black]
line-length = 100
target-version = ["py310", "py311", "py312"]
include = '\.pyi?$'
exclude = '''
/(
    \.git
  | \.mypy_cache
  | \.pytest_cache
  | \.venv
  | venv
  | build
  | dist
)/
'''

[tool.isort]
profile = "black"
line_length = 100
known_first_party = ["cap_srp"]

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
ignore_missing_imports = true
strict = true

[tool.pytest.ini_options]
minversion = "8.0"
addopts = "-ra -q"
testpaths = ["tests"]
asyncio_mode = "auto"

[tool.coverage.run]
source = ["cap_srp"]
branch = true

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
]

[tool.ruff]
line-length = 100
target-version = "py310"
select = ["E", "F", "W", "I", "N", "UP", "B", "C4"]
ignore = ["E501"]
